+++ src/Braintree/IUsBankAccountGateway.cs
@@ -1,0 +1,9 @@
+#pragma warning disable 1591
+
+namespace Braintree
+{
+    public interface IUsBankAccountGateway
+    {
+        UsBankAccount Find(string token);
+    }
+}