+++ test/Braintree.TestUtil/MerchantAccountIDs.cs
@@ -8,5 +8,6 @@ namespace Braintree.TestUtil
         public static string THREE_D_SECURE_MERCHANT_ACCOUNT_ID = "three_d_secure_merchant_account";
         public static string FAKE_AMEX_DIRECT_MERCHANT_ACCOUNT_ID = "fake_amex_direct_usd";
         public static string FAKE_VENMO_ACCOUNT_MERCHANT_ACCOUNT_ID = "fake_first_data_venmo_account";
+        public static string US_BANK_MERCHANT_ACCOUNT_ID = "us_bank_merchant_account";
     }
 }